requirements.txt
Flask==2.3.2
Flask_Limiter==3.3.1
PyYAML==6.0
requests==2.22.0
setuptools==65.5.0
waitress==2.1.2
Werkzeug==2.3.4
